dmaengine: sdxi: Make sdxi_copy->len u64, not size_t
Copy operations can be up to 4GB in size, so size_t isn't adequate on
32-bit architectures.

While 4GB transfers seem unlikely on such systems, this is necessary
to make the descriptor tests correct:

  drivers/dma/sdxi/descriptor_kunit.c: In function 'copy':
  include/uapi/linux/const.h:20:25: error: conversion from 'long long
    unsigned int' to 'size_t' {aka 'unsigned int'} changes value from
    '4294967297' to '1' [-Werror=overflow]
  ...
  drivers/dma/sdxi/descriptor_kunit.c:34:20: note: in expansion of
    macro 'SZ_4G'
     34 |         copy.len = SZ_4G + 1;
        |                    ^~~~~

+++ drivers/dma/sdxi/descriptor.h
@@ -55,7 +55,7 @@ static inline struct sdxi_cxt_range __sdxi_cxt_range(u16 a, u16 b)
 struct sdxi_copy {
 	dma_addr_t src;
 	dma_addr_t dst;
-	size_t len;
+	u64 len;
 	u16 src_akey;
 	u16 dst_akey;
 };
